+++ packages/contracts/contracts/core/component/MetaTxComponent.sol
@@ -40,7 +40,7 @@ abstract contract MetaTxComponent is Component, BaseRelayRecipient {
     function _msgSender()
         internal
         view
-        override(ContextUpgradeable, BaseRelayRecipient)
+        override(DaoAuthorizableUpgradeable, BaseRelayRecipient)
         returns (address)
     {
         return BaseRelayRecipient._msgSender();
@@ -50,7 +50,7 @@ abstract contract MetaTxComponent is Component, BaseRelayRecipient {
     function _msgData()
         internal
         view
-        override(ContextUpgradeable, BaseRelayRecipient)
+        override(DaoAuthorizableUpgradeable, BaseRelayRecipient)
         returns (bytes calldata)
     {
         return BaseRelayRecipient._msgData();