+++ packages/contracts/contracts/core/plugin/Plugin.sol
@@ -4,19 +4,19 @@ pragma solidity 0.8.10;
 
 import "@openzeppelin/contracts/utils/introspection/ERC165.sol";
 
-import {DaoAuthorizable} from "../component/DaoAuthorizable.sol";
+import {DaoAuthorizableConstructable} from "../component/DaoAuthorizableConstructable.sol";
 import {IDAO} from "../IDAO.sol";
 
 /// @title Plugin
 /// @author Aragon Association - 2022
-/// @notice An abstract, non-upgradeable contract to inherit from when creating a plugin being deployed via the `new` keyword.
-abstract contract Plugin is ERC165, DaoAuthorizable {
+/// @notice An abstract, non-upgradeDaoAuthorizableConstructableo inherit from when creating a plugin being deployed via the `new` keyword.
+abstract contract Plugin is ERC165, DaoAuthorizableConstructable {
     /// @notice The [ERC-165](https://eips.ethereum.org/EIPS/eip-165) interface ID of the contract.
     bytes4 public constant PLUGIN_INTERFACE_ID = type(Plugin).interfaceId;
 
     /// @notice Constructs the plugin by storing the associated DAO.
-    /// @param _dao The DAO contract.
-    constructor(IDAO _dao) DaoAuthorizable(_dao) {}
+    /// @param _dao The DAODaoAuthorizableConstructable
+    constructor(IDAO _dao) DaoAuthorizableConstructable(_dao) {}
 
     /// @notice Checks if this or the parent contract supports an interface by its ID.
     /// @param _interfaceId The ID of the interace.