+++ packages/contracts/contracts/plugin/examples/count/CounterV2.sol
@@ -56,8 +56,8 @@ contract CounterV2 is PluginUUPSUpgradeable {
 
     /// @notice Multiplies the count with a number.
     /// @param _a The number to multiply the coun with.
-    function multiply(uint256 _a) public auth(MULTIPLY_PERMISSION_ID) returns (uint256 count) {
-        count = multiplyHelper.multiply(count, _a);
+    function multiply(uint256 _a) public view auth(MULTIPLY_PERMISSION_ID) returns (uint256) {
+        return multiplyHelper.multiply(count, _a);
     }
 
     /// @notice Executes something on the DAO.